+++ src/main/java/miyucomics/hexical/mixin/BlockAkashicBookshelfMixin.java
@@ -1,12 +1,10 @@
 package miyucomics.hexical.mixin;
 
-import at.petrak.hexcasting.api.casting.eval.vm.CastingImage;
-import at.petrak.hexcasting.api.casting.iota.Iota;
 import at.petrak.hexcasting.api.casting.iota.IotaType;
 import at.petrak.hexcasting.common.blocks.akashic.BlockAkashicBookshelf;
 import at.petrak.hexcasting.common.blocks.akashic.BlockEntityAkashicBookshelf;
-import at.petrak.hexcasting.xplat.IXplatAbstractions;
 import miyucomics.hexical.inits.HexicalSounds;
+import miyucomics.hexical.utils.CastingUtils;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.entity.player.PlayerEntity;
@@ -22,8 +20,6 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-import java.util.List;
 
 import static net.minecraft.sound.SoundCategory.BLOCKS;
 
@@ -42,14 +38,7 @@ public class BlockAkashicBookshelfMixin {
 		if (nbt == null)
 			return;
 
-		CastingImage newImage = IXplatAbstractions.INSTANCE.getStaffcastVM((ServerPlayerEntity) player, hand).getImage();
-		Iota iota = IotaType.deserialize(nbt, (ServerWorld) world);
-		if (newImage.getParenCount() == 0)
-			newImage.getStack().add(iota);
-		else
-			newImage.getParenthesized().add(new CastingImage.ParenthesizedIota(iota, false));
-
-		IXplatAbstractions.INSTANCE.setStaffcastImage((ServerPlayerEntity) player, newImage);
+		CastingUtils.giveIota((ServerPlayerEntity) player, IotaType.deserialize(nbt, (ServerWorld) world));
 		world.playSound(null, pos, HexicalSounds.SUDDEN_REALIZATION, BLOCKS, 1f, 1f);
 		player.swingHand(hand, true);
 	}