+++ src/main/java/miyucomics/hexical/mixin/GameRendererMixin.java
@@ -1,0 +1,16 @@
+package miyucomics.hexical.mixin;
+
+import miyucomics.hexical.client.ShaderRenderer;
+import net.minecraft.client.render.GameRenderer;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(GameRenderer.class)
+public class GameRendererMixin {
+	@Inject(method = "render", at = @At("HEAD"))
+	public void renderShader(float deltaTick, long l, boolean bl, CallbackInfo ci) {
+		ShaderRenderer.render(deltaTick);
+	}
+}