+++ src/main/java/miyucomics/hexical/prestidigitation/AkashicCopyEffect.kt
@@ -5,6 +5,7 @@ import at.petrak.hexcasting.api.casting.iota.IotaType
 import at.petrak.hexcasting.common.blocks.akashic.BlockEntityAkashicBookshelf
 import at.petrak.hexcasting.xplat.IXplatAbstractions
 import miyucomics.hexical.interfaces.PrestidigitationEffect
+import miyucomics.hexical.utils.CastingUtils
 import net.minecraft.entity.Entity
 import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.util.Hand
@@ -18,9 +19,7 @@ class AkashicCopyEffect : PrestidigitationEffect {
 		val shelf = env.world.getBlockEntity(position)
 		if (shelf is BlockEntityAkashicBookshelf) {
 			val nbt = shelf.iotaTag ?: return
-			val vm = IXplatAbstractions.INSTANCE.getStaffcastVM(caster, Hand.MAIN_HAND)
-			vm.image.stack.toMutableList().add(IotaType.deserialize(nbt, caster.serverWorld))
-			IXplatAbstractions.INSTANCE.setStaffcastImage(caster, vm.image)
+			CastingUtils.giveIota(caster, IotaType.deserialize(nbt, caster.serverWorld))
 		}
 	}